+++ mrv2/docs/HISTORY.md
@@ -37,6 +37,7 @@ v0.7.2
 - Added Edit mode to the sessions file.
 - Added Edit mode to the network connection (it will load as timeline or full).
 - Laser annotations are no longer added to the draw undo/redo queue.
+- Laser annotations now work properly on Network connections.
 
 
 v0.7.1